Readme and unit test

+++ src/main/java/com/preassignment/musinsa/display/service/BrandService.java
@@ -3,16 +3,19 @@ package com.preassignment.musinsa.display.service;
 import com.preassignment.musinsa.display.dto.BrandRequestDTO;
 import com.preassignment.musinsa.display.entity.Brand;
 import com.preassignment.musinsa.display.repository.BrandRepository;
+import com.preassignment.musinsa.display.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class BrandService {
 
+  private final ProductRepository productRepository;
   private final BrandRepository brandRepository;
 
   @Autowired
-  public BrandService(BrandRepository brandRepository) {
+  public BrandService(ProductRepository productRepository, BrandRepository brandRepository) {
+    this.productRepository = productRepository;
     this.brandRepository = brandRepository;
   }
 
@@ -26,6 +29,12 @@ public class BrandService {
   public void updateBrand(Long id, BrandRequestDTO brandRequest) {
     Brand brand = brandRepository.findById(id)
         .orElseThrow(() -> new RuntimeException("해당 브랜드를 찾을 수 없습니다."));
+
+    // 브랜드의 카테고리에 최소 하나의 상품이 남는지 확인
+    if (productRepository.countByBrand(brand) == 1) {
+      throw new RuntimeException("브랜드의 카테고리에는 최소 하나의 상품이 존재해야 합니다.");
+    }
+
     brand.setName(brandRequest.getName());
     brandRepository.save(brand);
   }
@@ -34,6 +43,12 @@ public class BrandService {
   public void deleteBrand(Long id) {
     Brand brand = brandRepository.findById(id)
         .orElseThrow(() -> new RuntimeException("해당 브랜드를 찾을 수 없습니다."));
+
+    // 브랜드의 카테고리에 최소 하나의 상품이 남는지 확인
+    if (productRepository.countByBrand(brand) == 1) {
+      throw new RuntimeException("브랜드의 카테고리에는 최소 하나의 상품이 존재해야 합니다.");
+    }
+
     brandRepository.delete(brand);
   }
 }
